backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Database & Search
opensearch-py==2.4.2
requests-aws4auth==1.2.3

# File Processing
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Environment & Configuration
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1

# Jira Integration
jira==3.6.0

# Utilities
python-dateutil==2.8.2
